ApiGateway/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose==3.3.0
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.1
